reduced nn and fixed missing dir error

+++ player.py
@@ -1,3 +1,4 @@
+import os
 import numpy as np
 import environment
 import config
@@ -69,7 +70,8 @@ class Agent():
         # self.target_nn.load_dir("main_nn")
         # self.target_nn.save_model("target_nn", self.to_weights)
         self.target_nn = copy(self.main_nn)
-        rmtree(files.get_path("training/target_nn"))
+        if os.path.exists(files.get_path("training/target_nn")):
+            rmtree(files.get_path("training/target_nn"))
         copytree(files.get_path("training/main_nn"), files.get_path("training/target_nn"))
 
         files.edit_key("save.json", ["target_nn_version"], [self.main_nn.version])
